+++ RELEASE_NOTES.md
@@ -1,7 +1,15 @@
 ### New in 0.28 (not released yet)
 
- * Fixed: Prevent `abstract ICommandHandler<,,,>` from being registered in 
-   `EventFlowOptionsCommandHandlerExtensions.AddCommandHandlers`
+* Fixed: Prevent `abstract ICommandHandler<,,,>` from being registered in 
+   `EventFlowOptionsCommandHandlerExtensions.AddCommandHandlers(...)`
+* Fixed: Prevent `abstract IEventUpgrader<,>` from being registered in 
+   `EventFlowOptionsEventUpgradersExtensions.AddEventUpgraders(...)`
+* Fixed: Prevent `abstract IMetadataProvider` from being registered in 
+   `EventFlowOptionsMetadataProvidersExtensions.AddMetadataProviders(...)`
+* Fixed: Prevent `abstract IQueryHandler<,>` from being registered in 
+   `EventFlowOptionsQueriesExtensions.AddQueryHandlers(...)`
+* Fixed: Prevent `abstract ISubscribeSynchronousTo<,,>` from being registered in 
+   `EventFlowOptionsSubscriberExtensions.AddSubscribers(...)`
 
 ### New in 0.27.1765 (released 2016-02-25)